requirements.txt
streamlit>=1.37.0
pandas>=2.1.0
numpy>=1.26.0
matplotlib>=3.8.0
seaborn>=0.13.0
python-dotenv>=1.0.0
google-generativeai>=0.5.2
